Restructure /main folder
