+++ .expo/packager-info.json
@@ -1,10 +1,10 @@
 {
   "webpackServerPort": null,
-  "expoServerNgrokUrl": null,
-  "packagerNgrokUrl": null,
-  "ngrokPid": null,
+  "expoServerNgrokUrl": "https://fn-29n.minhchinh97h.tower-client.exp.direct",
+  "packagerNgrokUrl": "https://packager.fn-29n.minhchinh97h.tower-client.exp.direct",
+  "ngrokPid": 14708,
   "devToolsPort": 19002,
-  "expoServerPort": null,
-  "packagerPort": null,
-  "packagerPid": null
+  "expoServerPort": 19000,
+  "packagerPort": 19001,
+  "packagerPid": 16228
 }
