+++ App.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import Login from './components/login/Login' //Login screen
 import SignUp from './components/signup/SignUp' //Sign Up screen
-import MainNavigator from './components/main/Main' //Main screen
+import MainNavigator from './components/main' //Main screen
 import {createStackNavigator, createAppContainer, createSwitchNavigator} from 'react-navigation'
 import {createStore} from 'redux'
 import {Provider} from 'react-redux'
