+++ components/main/screens/Reward/Reward.js
